Move fixture bundling into a separate plugin.
This way, people can use it optionally if they have horrible, interdependent tests and deserve to be locked away behind a cellar door.

We manually add the fixture-bundling plugin to Nose's plugin manager (though it's disabled by default) so it's still available even if the setup.py entrypoint isn't around--for example, if django_nose is in a vendor lib.