Make FastFixtureTestCase and TransactionTestCaseReorderer get along better.
I imagined a situation in which the fixture bundler would shove a TransactionTestCase in the middle of a bundle of FFTCs and the TTC wouldn't clean up after itself, thus wrecking the bundle. So I made the bundler much more conservative: now it includes only FFTCs in bundles, which is fine, because nobody else was going to implement my advisory bits anyway.

Then the test types that concern the bundler and the TTC reorderer are disjunct, making the interaction of the 2 plugins much easier to think about.

I also narrowed the states returned by the TTC reorderer to 2, since all we really care about are "clean" and "dirty".