Improve doc string of runner.BasicNoseRunner.run_tests()
The first argument MAY NOT be just a string, as previously shown in
the doc string, but MUST be a sequence of strings.  This argument
is used as an argument to list(), which provides a list of the
characters in the string, not what you want.

Also add comment documenting the fact that the extra_tests argument is
currently ignored, see also issue jbalogh django-nose issue 47.