Tune output capturing in runtests.sh
In runtests.sh, redirect stdout and stderr to temporary files, and
print the captured output on errors rather than rerunning the tests.

Add a -v option to runtests.sh, to display test output while the tests
are running, instead of only on errors.

Adjust TravisCI and tox configuration to display verbose output in
TravisCI runs, and not run coveralls locally.