Don't reinitialize a TEST_MIRROR (duplicate work) on REUSE_DB=1.
When using REUSE_DB=1 and a setup with more than one database, where the
other databases point to the first one, using the TEST_MIRROR setting,
Nose would sqlflush both databases.

When there is lots of work or a slow link, running ALTER TABLE to reset
the auto increment on all tables can take more than a few seconds.
Before this fix, the whole process would be done a second time on the
test-mirror database -- which is the same.