Allow FastFixtureTestCase without fixtures
In the scenario that a test file does something like::

    from django_nose.testcases import FastFixtureTestCase as TestCase

And then has a bunch of fixture related test cases but *one* test case
that doesn't have fixtures - because why import another test class for
that one little test case? - in that scenario the testcase will raise
an error iterating over a NoneType. Ensures that fixtures are only
handled when actually present.